requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.6.2
apache-airflow==2.6.1
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==5.2.2
appnope==0.1.3
argcomplete==3.0.8
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.7.3
black==23.3.0
bleach==6.0.0
blinker==1.6.2
boto3==1.26.133
botocore==1.29.133
cachelib==0.9.0
cattrs==22.2.0
cbor2==5.4.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
cmdstanpy==1.1.0
colorama==0.4.6
colorlog==4.8.0
comm==0.1.3
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.0.7
convertdate==2.4.0
cron-descriptor==1.3.0
croniter==1.3.14
cryptography==40.0.2
crytic-compile==0.3.1
cycler==0.11.0
cytoolz==0.12.1
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.20.1
elastic-transport==8.4.0
elasticsearch==8.7.0
email-validator==1.3.1
ephem==4.1.4
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
executing==1.2.0
findspark==2.0.1
Flask==2.2.5
Flask-AppBuilder==4.3.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.39.4
frozenlist==1.3.3
gensim==4.3.1
gensim-sum-ext==0.1.2
graphviz==0.20.1
gunicorn==20.1.0
h11==0.14.0
hexbytes==0.3.0
holidays==0.25
httpcore==0.17.1
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
install==1.3.5
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
keyring==23.13.1
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
lru-dict==1.1.8
LunarCalendar==0.0.9
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.3
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
parsimonious==0.9.0
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.0.0
prettytable==3.7.0
prison==0.2.1
prompt-toolkit==3.0.38
prophet==1.1.3
protobuf==4.23.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pycorenlp==0.3.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.7.0
PyMeeus==0.5.12
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.4.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.0.2
readme-renderer==37.3
regex==2023.5.5
requests==2.30.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rich==13.3.5
rich_argparse==1.1.0
rlp==3.0.0
s3transfer==0.6.1
scipy==1.10.1
setproctitle==1.3.2
six==1.16.0
slither-analyzer==0.9.3
smart-open==6.3.0
sniffio==1.3.0
solc-select==1.0.3
soupsieve==2.4.1
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
textblob==0.17.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
twine==4.0.2
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.15
wcwidth==0.2.6
web3==6.4.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.2.3
wikipedia==1.4.0
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
